lib: lower WorkspaceCommandEnvironment to new module workspace_util
This also removes all `&Ui` from all its (re)load paths. Its the final step so it can get moved into the library.
